Try solving many of the test errors by adding init files
