+++ tests/base/pipeline/test_pipe.py
@@ -4,7 +4,7 @@ from pytest import fixture
 
 from pype.base.data.dataset import DataSet
 from pype.base.pipeline.pipe import Pipe
-from tests.test_utils import pytest_assert
+from tests.utils import pytest_assert
 
 
 class TestPipe:
