+++ tests/utils.py
@@ -1,22 +1,15 @@
-import shutil
 from contextlib import contextmanager
 from pathlib import Path
-from typing import Iterable
-
-from pytest import fixture
 
 from pype.base.data.data_source import DataSource
 from pype.base.data.dataset_source import DataSetSource
 from pype.base.evaluate.evaluator import Evaluator
-from pype.base.experiment.experiment import Experiment
-from pype.base.logger.local_logger import LocalLogger
 from pype.base.model import Model
 from pype.base.pipeline.operator import Operator
 from pype.base.pipeline.pipe import Pipe
 from pype.base.pipeline.pipeline import Pipeline
 from pype.base.pipeline.type_checker import DataModel, TypeChecker, TypeCheckerPipe
-from pype.base.serialiser.joblib_serialiser import JoblibSerialiser
-from tests.test_utils_training_support import reverse
+from tests.utils_training_support import reverse
 
 
 class AnyArg:
@@ -126,29 +119,3 @@ def get_dummy_type_checkers() -> tuple[TypeCheckerPipe, TypeCheckerPipe]:
     return TypeCheckerPipe("input", ["x"], type_checker_classes=[(list, DummyTypeChecker)]), TypeCheckerPipe(
         "output", ["y"], type_checker_classes=[(list, DummyTypeChecker)]
     )
-
-
-@fixture(scope="module")
-def dummy_experiment() -> Iterable[Experiment]:
-    train = get_dummy_data(20, 1, 0)
-    test = get_dummy_data(5, 2, -1)
-
-    output_folder = Path(__file__).parent / "tmp_folder"
-
-    input_checker, output_checker = get_dummy_type_checkers()
-
-    yield Experiment(
-        data_sources={"train": train, "test": test},
-        model=DummyModel(inputs=["x"], outputs=["y"]),
-        pipeline=get_dummy_pipeline(),
-        logger=LocalLogger(),
-        evaluator=get_dummy_evaluator(),
-        serialiser=JoblibSerialiser(),
-        output_folder=output_folder,
-        input_type_checker=input_checker,
-        output_type_checker=output_checker,
-        parameters={"param_1": 2},
-    )
-
-    # make sure we purge the output folder again after a test is done.
-    shutil.rmtree(output_folder)
